requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
qrcode>=7.4.2
pillow>=10.0.0
